[FE] add : post data fetching
