+++ src/app/page.tsx
@@ -6,7 +6,9 @@ export default function Home() {
   return (
     <section className="w-full">
       <Profile />
+      {/* @ts-expect-error Server Component */}
       <FeaturedPosts />
+      {/* @ts-expect-error Server Component */}
       <YouMayLike />
     </section>
   );
