+++ src/components/Card.tsx
@@ -1,15 +1,20 @@
 import Image from "next/image";
 import reviewImage from "../../public/images/posts/review-2023.png";
+import { Post } from "@/app/api/posts";
 
-const Card = () => {
+interface CardProps {
+  post: Post;
+}
+
+const Card = ({ post }: CardProps) => {
   return (
     <div className="rounded border-2 border-gray-300">
       <Image src={reviewImage} alt="thumbnail" />
-      <div className="text-right">날짜</div>
+      <div className="text-right">{post.date}</div>
       <div className="flex flex-col items-center">
-        <div>제목</div>
-        <div>부제목</div>
-        <div>태그</div>
+        <div>{post.title}</div>
+        <div>{post.description}</div>
+        <div>{post.category}</div>
       </div>
     </div>
   );
