+++ src/components/FeaturedPosts.tsx
@@ -1,16 +1,15 @@
-import { getFeaturedPosts } from "@/app/api/posts";
+import { Post, getFeaturedPosts } from "@/app/api/posts";
 import Card from "./Card";
 
-const FeaturedPosts = () => {
-  const featuredPosts = getFeaturedPosts();
+const FeaturedPosts = async () => {
+  const featuredPosts = await getFeaturedPosts();
   return (
     <div className="max-w-screen-xl mx-auto">
       <h3 className="font-bold">Featured Posts</h3>
       <div className="grid grid-cols-3 gap-4">
-        <Card />
-        <Card />
-        <Card />
-        <Card />
+        {featuredPosts.map((post: Post, idx) => {
+          return <Card key={idx} post={post} />;
+        })}
       </div>
     </div>
   );
