+++ src/components/YouMayLike.tsx
@@ -1,40 +1,18 @@
-"use client";
+import { getNonFeaturedPosts } from "@/app/api/posts";
 import Card from "@/components/Card";
+import MultiCarousel from "./MultiCarousel";
 
-import Carousel from "react-multi-carousel";
-import "react-multi-carousel/lib/styles.css";
+const YouMayLike = async () => {
+  const nonFeaturedPosts = await getNonFeaturedPosts();
 
-const YouMayLike = () => {
-  const responsive = {
-    superLargeDesktop: {
-      // the naming can be any, depends on you.
-      breakpoint: { max: 4000, min: 3000 },
-      items: 5,
-    },
-    desktop: {
-      breakpoint: { max: 3000, min: 1024 },
-      items: 5,
-    },
-    tablet: {
-      breakpoint: { max: 1024, min: 464 },
-      items: 2,
-    },
-    mobile: {
-      breakpoint: { max: 464, min: 0 },
-      items: 1,
-    },
-  };
   return (
     <div className="max-w-screen-xl mx-auto">
       <h3 className="font-bold">You may like</h3>
-      <Carousel responsive={responsive}>
-        <Card />
-        <Card />
-        <Card />
-        <Card />
-        <Card />
-        <Card />
-      </Carousel>
+      <MultiCarousel>
+        {nonFeaturedPosts.map((post, idx) => {
+          return <Card key={idx} post={post} />;
+        })}
+      </MultiCarousel>
       ;
     </div>
   );
